rename channels.py to listchannels.py
